requirements.txt
browser-use
langchain-openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
selenium
html5lib 
lxml